+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Erich Matos Viegas: fixed chat messages
 - Erich Matos Viegas: you can custom everything
 - Settings Refactor 
+- FilePicker works on the forge
+- default dice color 
 
 # v1.9
 - v13